[Smoke-tests] Move tests into a single package.
This changes places all of the tests into a single package. There is no
reason to separate them and have one file per package.

This change also renames AllTests to TestSuite, as that better
identifies its purpose.
