+++ smoke-tests/src/main/java/com/google/firebase/testing/DynamicLinksTest.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.firebase.testing.dynamiclinks;
+package com.google.firebase.testing;
 
 import static com.google.common.truth.Truth.assertThat;
